lyra/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
lxml>=5.2.0
httpx>=0.27.0
gspread>=6.1.0
google-auth>=2.29.0
python-telegram-bot>=21.0
vcrpy>=6.0.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
respx>=0.21.0
ruff>=0.4.0
black>=24.4.0
mypy>=1.10.0
cachetools>=5.3.0
aiofiles>=23.2.0
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
